Create test_flatten_pipeline_metadata.py
