+++ tests/non_search/core/test_pipeline.py
@@ -4,10 +4,7 @@ from unittest.mock import patch
 
 from click.testing import CliRunner
 
-from app.repository.pipeline import (
-    _flatten_pipeline_metadata,
-    generate_pipeline_ingest_input,
-)
+from app.repository.pipeline import generate_pipeline_ingest_input
 from app.service.pipeline import format_pipeline_ingest_input, get_db_state_content
 from scripts.db_state_validator_click import main as db_state_validator_main
 from tests.non_search.setup_helpers import (
@@ -76,18 +73,6 @@ def test_generate_pipeline_ingest_input__deleted(data_db):
     assert len(documents) == 1
     assert documents[0].name == "Fam1"
     assert documents[0].import_id == "CCLW.executive.1.2"
-
-
-def test_flatten_pipeline_metadata():
-    family_metadata = {"a": ["1"], "b": ["2"]}
-    doc_metadata = {"a": ["3"], "b": ["4"]}
-    result = _flatten_pipeline_metadata(family_metadata, doc_metadata)
-
-    assert len(result) == 4
-    assert result["family.a"] == ["1"]
-    assert result["family.b"] == ["2"]
-    assert result["document.a"] == ["3"]
-    assert result["document.b"] == ["4"]
 
 
 def test_get_db_state_content_success(data_db, caplog):
